init refacotr
