--- PlantApp/Components/ScrollSection.swift
@@ -1,71 +1,0 @@
-//
-//  ScrollSection.swift
-//  PlantCodeiOS15
-//
-//  Created by Aeliana Shen on 6/14/23.
-//
-
-import SwiftUI
-
-struct ScrollSection: View {
-    private let plantService = PlantService()
-    @State private var plantList: [Plant]?
-    @State private var showingError = false
-    @State private var errorMessage = ""
-    
-    var body: some View {
-        VStack (alignment: .leading) {
-            Text("Recommended list for you")
-                .font(.headline)
-                .foregroundColor(ProjColor.Gumleaf)
-                .padding(.horizontal, 25.0)
-            
-            ScrollView(.horizontal, showsIndicators: false){
-                HStack {
-                    ForEach(plantList ?? [], id: \.productID) { product in
-                        NavigationLink {
-                            ProductDetailView(plant: product)
-                        } label: {
-                            PlantCard(plant: product)
-                        }
-                    }
-                    if (plantList == nil) {
-                        Text("The recommended list has no data.")
-                            .font(.headline)
-                            .foregroundColor(ProjColor.PrimaryGreen)
-                    }
-                }
-                .padding(50)
-            }
-        }
-        .task {
-            do {
-                plantList = try await plantService.getPlants()
-                // print(try await plantService.getPlant(withId: "100002"))
-            } catch URLError.badURL {
-                showingError = true
-                errorMessage = "bad URL"
-            } catch URLError.badServerResponse {
-                showingError = true
-                errorMessage = "bad server response"
-            } catch URLError.cannotDecodeContentData {
-                showingError = true
-                errorMessage = "cannot decode content data"
-            } catch {
-                showingError = true
-                errorMessage = "unexpected error"
-            }
-        }
-        .alert(isPresented: $showingError, content: {
-            Alert(title: Text("Load the local plant list?"), message: Text("Problems occur when loading the recommended list from the internet. Click OK to load the local plant list. \n\n Error: \(errorMessage)"), primaryButton: .default(Text("Load")) {
-                plantList = Plant.localPlantList
-            }, secondaryButton: .cancel())
-        })
-    }
-}
-
-struct ScrollSection_Previews: PreviewProvider {
-    static var previews: some View {
-        ScrollSection()
-    }
-}
